+++ spec/integration/resource/api300/synergy/switch/create_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API300::C7000::Switch
+klass = OneviewSDK::API300::Synergy::Switch
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
   include_context 'integration api300 context'