+++ CHANGELOG.md
@@ -7,13 +7,12 @@
 #### Bug fixes & Enhancements:
 - [#201](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/201) Code to search the collection of resources (paginated search) is repeated in some resources
 - [#119](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/112) VolumeAttachment::remove_extra_unmanaged_volume throw Unexpected Http Error
+- [#202](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/202) The method #get_default_settings in LogicalInterconnectGroup is used on integration test
 
 #### Design changes:
    - Architecture for future API500 support. Features for API500 are not yet supported.
 
 # v4.1.0
-#### Bug fixes & Enhancements:
-- [#202](https://github.com/HewlettPackard/oneview-sdk-ruby/issues/202) The method #get_default_settings in LogicalInterconnectGroup is used on integration test
 
 #### New Resources:
 Added full support to Image Streamer Rest API version 300: