+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased Changes
-## Suggested release: v5.0.0
+## Suggested release: v4.1.1
 
 #### New Resources:
 (none)